detector/requirements.txt
flask==2.0.3
werkzeug==2.0.3
torch==2.0.0
transformers==4.34.0
pillow==8.3.1
opencv-python==4.5.3.56
numpy==1.21.2
requests==2.26.0
